+++ lib/base/Mem.h
@@ -154,7 +154,7 @@ void BaseNumaFreePages(void *ptr);
 /********************Interface NUMA**************************************/
 int numa_node_to_cpusmask(int node, uint64_t *cpusmask, int *nr);
 void msf_numa_tonode_memory(void *mem, size_t size, int node);
-long msf_numa_node_size(int node, long long * freep);
+long long msf_numa_node_size(int node, long long * freep);
 int msf_numa_distance(int node1, int node2);
 
 } /**************************** end namespace BASE ****************************/
